Fix oneAPI kernel compilation after recent refactor
Didn't notice such breaking changes while working on the build system.

This is a minimum possible change which seems to pass compilation of the
oneAPI kernel. Testing on the actual device would take a while.

More proper approach seems to wrap those fields into a structure within
the global data, so that no prefixing of the name is needed.