Cycles: Focus on JIT oneAPI kernels for Windows
Disable AoT compilation on Windows and disable oneAPI on Linux.

The former is due to the buildbot taking 80+ minutes to perform
AoT kernel compilation. The latter one is due to an issue somewhere
in the tech stack which leads to crash on runtime due to ABI
incompatibility.

This means that the first deliverable would be Windows support
with JIT compilation (on artists side).