Move the get_uv_map_vert/edge/pin_selection_name() functions to C from C++ and avoid std::string.
We need these functions both in C and in C code, also the std::string
concatenation was quite slow because it onvolves allocation. It now
uses a (stack allocated) char buffer passed in from the caller.

This alos gives us a single place in the code to handle overflowing
MAX_CUSTOMDATA_LAYER_NAME (to be done in a future commit).