Change MAX_CUSTOMDATA_LAYER_NAME to 72 characters.
We need four extra characters compared to the maximum name length the
user is allowed to provide, to store the prefixes for associated
layers. Changing MAX_CUSTOMDATA_LAYER_NAME to 68 leads to lots of dna
alignment errors which need padding. So instead of padding we just
increase the length to 72.