Delete an erroneously created UV attribute .
When loading a mesh without UVs from a wavefront obj file which also contains
meshes with UVs, we ended up creating an uninitialized UV layer. Detect when
that happens and delete the invaliud layer. We could also check for UVs
before creating the layer, but as this is probably a rather rare occasion
and checking involves looping over the whole mesh we just detect it after the
fact and delete.