Added test and fix for a rare situation with degenerate tets.
When degenerate tet is sandwiched between valid (non-degenerate) tets
it could be that a new point intersects the circumsphere on either side
of the degenerate tet. Degenerate tets were returning false in the
Delaunay check, effectively considered containing no points. The result
is that when replacing the tets on either side of the degenerate section
they try to insert the same edge twice into the boundary edge map, which
should never happen.

We have to consider a degenerate tet to have infinte circumradius and
contain all the points. That way degenerate triangles will always be
replaced.